requirements.txt
fastapi==0.75.0
redis-om==0.0.20
uvicorn==0.17.6
requests==2.27.1
strawberry-graphql[fastapi]